add mac build
