+++ .github/workflows/pull-request-linux.yml
@@ -6,9 +6,6 @@ on:
       - "**/Cargo.lock"
       - "**/Cargo.toml"
       - "**/package.json"
-      - "**.svelte"
-      - "**.js"
-      - "**.ts"
     branches:
       - main
 env:
