+++ .github/workflows/pull-request-macos.yml
@@ -1,0 +1,22 @@
+name: On Pull Request To Main (MacOS)
+on:
+  pull_request:
+    paths:
+      - "**.rs"
+      - "**/Cargo.lock"
+      - "**/Cargo.toml"
+      - "**/package.json"
+    branches:
+      - main
+env:
+  CLICOLOR_FORCE: 1
+jobs:
+  pre-merge:
+    name: Pre-merge checks
+    runs-on: macos-latest
+    steps:
+      - name: Checkout Repository
+        uses: actions/checkout@v3
+      - uses: extractions/setup-just@v1
+      - name: Build player
+        run: just build-player aarch64-apple-darwin
