+++ .github/workflows/release-macos.yml
@@ -1,0 +1,27 @@
+name: Create release, build and upload artifacts (MacOS)
+
+permissions:
+  contents: write
+
+on:
+  push:
+    tags:
+      - v[0-9]+.*
+env:
+  CLICOLOR_FORCE: 1
+jobs:
+  create-build-upload:
+    name: Build, package and publish
+    runs-on: macos-latest
+    steps:
+      - name: Checkout Repository
+        uses: actions/checkout@v3
+      - uses: extractions/setup-just@v1
+      - name: Build player
+        run: just build-player aarch64-apple-darwin
+      - name: Tar qobuz-player-aarch64-apple-darwin.tar.gz
+        run: tar -cvf qobuz-player-aarch64-apple-darwin.tar.gz -C ./target/aarch64-apple-darwin/release/ qobuz-player
+      - name: Release ${{ matrix.target }}
+        uses: softprops/action-gh-release@v1
+        with:
+          files: qobuz-player-aarch64-apple-darwin.tar.gz
